\name{aRtsy-package}
\alias{aRtsy}
\alias{aRtsy-package}
\docType{package}
\title{aRtsy --- Generative Art using \code{ggplot2}}
\description{
  \if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

  \code{aRtsy} is an attempt at making generative art available for the masses in a simple and standardized format. 
  The package provides various algorithms for creating artworks in ggplot2 that incorporate some form of 
  randomness (depending on the set \code{seed}). Each type of artwork is implemented in a separate function.

  For documentation on \code{aRtsy} itself, including the manual and user guide
  for the package, worked examples, and other
  tutorial information visit the \href{https://koenderks.github.io/aRtsy/}{package website}.
  
}

\author{
  \tabular{ll}{
    Koen Derks (maintainer, author)\tab <koen-derks@hotmail.com> \cr
  }
 
  Please use the citation provided by \R when citing this package. 
  A BibTex entry is available from \code{citation("aRtsy")}.
}

\seealso{
Useful links:
\itemize{
  \item The \href{https://twitter.com/aRtsy_package}{twitter feed} to check the artwork of the day.
  \item The \href{https://github.com/koenderks/aRtsy/issues}{issue page} to submit a bug report or feature request.
}
}

\keyword{aRtsy} 
\keyword{package}