% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_flame.R
\name{canvas_flame}
\alias{canvas_flame}
\title{Draw a Fractal Flame}
\usage{
canvas_flame(colors, background = "#000000",
                      iterations = 1000000, resolution = 1000,
                      blend = TRUE, post = FALSE, final = FALSE)
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{background}{a character specifying the color used for the background.}

\item{iterations}{a positive integer specifying the number of iterations of the algorithm.}

\item{resolution}{resolution of the artwork in pixels per row/column. Increasing the resolution increases the quality of the artwork but also increases the computation time exponentially.}

\item{blend}{logical. Whether to blend the variations (\code{TRUE}) or pick a unique variation in each iteration (\code{FALSE}).}

\item{post}{logical. Whether to apply a posttransformation in each iteration.}

\item{final}{logical. Whether to apply a final transformation in each iteration.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function implements the fractal flame algorithm.
}
\examples{
\donttest{
set.seed(2)

# Simple example
canvas_flame(colors = colorPalette("origami"))
}

}
\references{
\url{https://en.wikipedia.org/wiki/Fractal_flame}
}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
