% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_flow.R
\name{canvas_flows}
\alias{canvas_flows}
\title{Paint Flows on a Canvas}
\usage{
canvas_flows(colors, background = "#fafafa", lines = 100,
             iterations = 500, width = 100, height = 100)
}
\arguments{
\item{colors}{a string specifying the color used for the artwork.}

\item{background}{a character specifying the color used for the background.}

\item{lines}{the number of lines to draw.}

\item{iterations}{the maximum number of iterations for each line.}

\item{width}{a positive integer specifying the width of the artwork in pixels.}

\item{height}{a positive integer specifying the height of the artwork in pixels.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function paints watercolors on a canvas.
}
\examples{
\donttest{
set.seed(10)
canvas_flows(colors = colorPalette("tuscany1"), lines = 500)
}

}
\references{
\url{https://tylerxhobbs.com/essays/2020/flow-fields}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
