% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_function.R
\name{canvas_function}
\alias{canvas_function}
\title{Paint A Random Function on a Canvas}
\usage{
canvas_function(color, background = '#fafafa')
}
\arguments{
\item{color}{a string specifying the color used for the artwork.}

\item{background}{a character specifying the color used for the background.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function paints functions with random parameters and mimics the functionality of the \code{generativeart} package.
}
\examples{
\donttest{
set.seed(1)

# Simple example
canvas_function(color = 'navyblue')
}

}
\references{
\url{https://github.com/cutterkom/generativeart}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
