% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_gemstone.R
\name{canvas_gemstone}
\alias{canvas_gemstone}
\title{Paint a gemstone on a canvas}
\usage{
canvas_gemstone(colors, maxk = 1, n = 1000, resolution = 500)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the artwork.}

\item{maxk}{the maximum number of nearest neighbors to consider.}

\item{n}{number of data points to generate.}

\item{resolution}{the number of pixels (width and height) of the artwork.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function creates an artwork from randomly generated data by running a k-nearest neighbors regression algorithm to predict the color of each pixel on the canvas.
}
\examples{
\donttest{
set.seed(1)
canvas_gemstone(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))
}

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
