% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_mesh.R
\name{canvas_mesh}
\alias{canvas_mesh}
\title{Draw Meshes}
\usage{
canvas_mesh(colors, background = "#fafafa", 
             transform = c("perlin", "fbm", "simplex", "cubic",
                           "worley", "knn", "rf", "svm"),
             n = 301, iterations = 500, mixProb = 0)
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{background}{a character specifying the color used for the background (and the hole).}

\item{transform}{a character specifying the type of transformation to use for the radius.}

\item{n}{an integer specifying the number of lines.}

\item{iterations}{a positive integer specifying the number of iterations of the algorithm.}

\item{mixProb}{a value between 0 and 1 specifying the probability of a line segment having another color.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function draws meshes.
}
\examples{
\donttest{
set.seed(2)

# Simple example
canvas_mesh(colors = colorPalette("origami"))
}

}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
