% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_mosaic.R
\name{canvas_mosaic}
\alias{canvas_mosaic}
\title{Paint a mosaic on a canvas}
\usage{
canvas_mosaic(colors, maxk = 10, n = 1000, resolution = 500)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the artwork.}

\item{maxk}{the maximum number of nearest neighbors to consider.}

\item{n}{number of data points to generate.}

\item{resolution}{the number of pixels (width and height) of the artwork.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function paints a mosaic from randomly generated data by running a k-nearest neighbors classification algorithm to predict the color of each pixel on the canvas. Low values of \code{maxk} produce a mosaic like artwork, while higher values produce a more smooth decision boundary.
}
\examples{
\donttest{
set.seed(1)
canvas_mosaic(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))
}

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
