% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_petri.R
\name{canvas_petri}
\alias{canvas_petri}
\title{Draw Petri-dish Colonies}
\usage{
canvas_petri(colors, background = "#fafafa", circle = "black",
                attractors = 1000, hole = 0)
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{background}{a character specifying the color used for the background.}

\item{circle}{a character specifying the color used for the circle.}

\item{attractors}{a numeric value larger than 1000 specifying the number of attractors.}

\item{hole}{a value between 0 and 0.9 specifying the hole width.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function uses a space colony algorithm.
}
\examples{
\donttest{
set.seed(2)

# Simple example
canvas_petri(colors = colorPalette("origami"))

# Advanced example
canvas_petri(colors = "black", hole = 0.8)
}

}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
