% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_smoke.R
\name{canvas_smoke}
\alias{canvas_smoke}
\title{Draw Rainbow Smoke}
\usage{
canvas_smoke(resolution = 100, distance = c("minimum", "average"))
}
\arguments{
\item{resolution}{resolution of the artwork in pixels per row/column. Increasing the resolution increases the quality of the artwork but also increases the computation time exponentially.}

\item{distance}{an character specifying whether to take the neighbor with the minimum distance or the smallest average distance.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function creates rainbow smoke.
}
\examples{
\donttest{
set.seed(1)

# Simple example
canvas_smoke(n_colors = 100)
}

}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
