% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_splatter.R
\name{canvas_splatter}
\alias{canvas_splatter}
\title{Draw Paint Splatters}
\usage{
canvas_splatter(
  colors,
  background = "#fafafa",
  iterations = 100,
  n = 100,
  lwd = 0.1,
  resolution = 500
)
}
\arguments{
\item{colors}{a character (vector) specifying the color(s) used for the
artwork.}

\item{background}{a character specifying the color used for the background.}

\item{iterations}{a positive integer specifying the number of iterations of
the algorithm.}

\item{n}{a positive integer specifying the number of particles.}

\item{lwd}{expansion factor for the line width.}

\item{resolution}{resolution of the artwork in pixels per row/column.
Increasing the resolution increases the quality of the artwork but also
increases the computation time exponentially.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function draws paint splatters on a canvas using a particle
  system.
}
\examples{
\donttest{
set.seed(1)

# Simple example
canvas_splatter(colors = colorPalette("mixer1"))
}

}
\references{
\url{https://mattdesl.svbtle.com/generative-art-with-nodejs-and-canvas}
}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
