% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_stripes.R
\name{canvas_stripes}
\alias{canvas_stripes}
\title{Paint stripes on a canvas}
\usage{
canvas_stripes(colors, n = 300, H = 1, burnin = 1)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the artwork.}

\item{n}{the length of the brownian motion.}

\item{H}{the square of the standard deviation of each step.}

\item{burnin}{the number of brownian motion steps to discard before filling a row.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function creates a brownian motion on each row of the artwork and colors it according to the height of the motion.
}
\examples{
\donttest{
set.seed(1)
canvas_stripes(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))
}

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
