% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_watercolors.R
\name{canvas_watercolors}
\alias{canvas_watercolors}
\title{Paint Watercolors on a Canvas}
\usage{
canvas_watercolors(colors, background = "#fafafa", layers = 50,
                   depth = 1, width = 250, height = 250)
}
\arguments{
\item{colors}{a string specifying the color used for the artwork.}

\item{background}{a character specifying the color used for the background.}

\item{layers}{the number of layers of each color.}

\item{depth}{the maximum depth of the recusive algorithm.}

\item{width}{a positive integer specifying the width of the artwork in pixels.}

\item{height}{a positive integer specifying the height of the artwork in pixels.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function paints watercolors on a canvas.
}
\examples{
\donttest{
set.seed(10)
canvas_watercolors(colors = colorPalette("tuscany1"))
}

}
\references{
\url{https://tylerxhobbs.com/essays/2017/a-generative-approach-to-simulating-watercolor-paints}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
