% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPalette.R
\name{colorPalette}
\alias{colorPalette}
\title{Color palette generator.}
\usage{
colorPalette(name, n = NULL)
}
\arguments{
\item{name}{name of the color palette. See the \code{details} section for a list of implemented palettes.}

\item{n}{the number of colors to select from the palette. Required if \code{name = 'random'}. Otherwise, if \code{NULL}, automatically selects all colors from the chosen palette.}
}
\value{
A vector of colors.
}
\description{
This function creates a random color palette, or allows the user to select a pre-implemented palette.
}
\details{
The following color palettes are implemented:

\itemize{
 \item{\code{random}}
 \item{\code{tuscany1}}
 \item{\code{tuscany2}}
}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{canvas}
\keyword{palette}
