% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise.R
\name{noise}
\alias{noise}
\title{Generate Noise}
\usage{
noise(dims, k = 20, n = 100)
}
\arguments{
\item{dims}{a vector containing the two dimenstions of the output.}

\item{k}{a positive integer specifying the number of nearest neighbors to consider.}

\item{n}{a positive integer specifying the number of random data points to generate.}
}
\value{
A matrix containing the noise values.
}
\description{
This function generates noise. Currently it is only capable of generating k-nearest neighbors noise.
}
\examples{
\donttest{
set.seed(6)
plot(as.raster(noise(dims = c(100, 100)))
}

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{noise}
