% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_ant.R
\name{paint_ant}
\alias{paint_ant}
\title{Paint Langton's Ant}
\usage{
paint_ant(colors = '#000000', background = '#fafafa', seed = 1, 
          width = 200, height = 200)
}
\arguments{
\item{colors}{the color of the turmite.}

\item{background}{the color of the background.}

\item{seed}{the seed for the painting.}

\item{width}{the width of the painting.}

\item{height}{the height of the painting.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints Langton's Ant. Langton's ant is a two-dimensional universal Turing machine with a very simple set of rules but complex emergent behavior.
}
\examples{
paint_ant(colors = '#000000', background = '#fafafa', seed = 1,
          width = 200, height = 200)

}
\references{
\url{https://en.wikipedia.org/wiki/Langton\%27s_ant}
}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_shape}} \code{\link{paint_turmite}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
