% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_arcs.R
\name{paint_arcs}
\alias{paint_arcs}
\title{Paint Arcs on a Canvas}
\usage{
paint_arcs(colors, background = '#fdf5e6', n = 1, nrow = NULL, ncol = NULL, dir = 'right')
}
\arguments{
\item{background}{a character specifying the color used for the background.}

\item{n}{an integer specifying how many paintings should be put on the canvas.}

\item{nrow}{(optional) number of rows on the canvas}

\item{ncol}{(optional) number of columns on the canvas.}

\item{dir}{a character specifying which direction the arcs turn.}

\item{color}{a character specifying the 3 colors used for the arcs.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints arcs.
}
\examples{
paint_arcs(colors = c('black', 'red', 'yellow'))

}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_turmite}} \code{\link{paint_ant}} \code{\link{paint_mondriaan}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
