% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_circlemap.R
\name{paint_circlemap}
\alias{paint_circlemap}
\title{Paint a Circle Map on a Canvas}
\usage{
paint_circlemap(colors, x_min = 0, x_max = 12.56, y_min = 0, y_max = 1, 
               iterations = 10, width = 1500, height = 1500)
}
\arguments{
\item{colors}{a character specifying the color used for the function shape.}

\item{x_min}{a numeric value specifying the minimum value for the x-axis.}

\item{x_max}{a numeric value specifying the maximum value for the x-axis.}

\item{y_min}{a numeric value specifying the minimum value for the y-axis.}

\item{y_max}{a numeric value specifying the maximum value for the y-axis.}

\item{seed}{the seed for the painting.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function is my attempt at a circle map.
}
\examples{
paint_circlemap(colors = c('black', 'red', 'yellow'))

}
\references{
\url{https://linas.org/art-gallery/circle-map/circle-map.html}
}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_turmite}} \code{\link{paint_ant}} \code{\link{paint_mondriaan}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
