% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_diamond.R
\name{paint_diamonds}
\alias{paint_diamonds}
\title{Paint A Diamond on Canvas}
\usage{
paint_diamonds(colors, background = '#fafafa', col.line = 'black', 
                  radius = 10, alpha = 1, size = 0.25, p = 0.2, width = 500, height = 500)
}
\arguments{
\item{colors}{a character (vector) specifying the colors used for the strokes.}

\item{background}{a character specifying the color used for the borders.}

\item{col.line}{color of the lines.}

\item{radius}{radius of the diamonds.}

\item{alpha}{transparency of the diamonds If \code{NULL}, added layers become increasingly more transparent.}

\item{size}{size of the borders}

\item{p}{takeover probability.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function draws many points on the canvas and connects these points into a polygon. After repeating this for all the colors, the edges of all polygons are drawn on top of the painting.
}
\examples{
set.seed(1)
paint_diamonds(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'), radius = 10)

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
