% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_function.R
\name{paint_function}
\alias{paint_function}
\title{Paint Functions on a Canvas}
\usage{
paint_function(color, background = '#fafafa', seed = 1)
}
\arguments{
\item{color}{a character specifying the color used for the function shape.}

\item{background}{a character specifying the color used for the background.}

\item{seed}{the seed for the painting.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints functions and mimics the functionality of the \code{generativeart} package.
}
\examples{
paint_function(color = '#000000', background = '#fafafa')

}
\references{
\url{https://github.com/cutterkom/generativeart}
}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_turmite}} \code{\link{paint_ant}} \code{\link{paint_mondriaan}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
