% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_mondriaan.R
\name{paint_mondriaan}
\alias{paint_mondriaan}
\title{Paint a Mondriaan on a Canvas}
\usage{
paint_mondriaan(colors, background = '#000000', cuts = 50, ratio = 1.618,
                width = 100, height = 100)
}
\arguments{
\item{colors}{a character vector specifying the colors used in the squares.}

\item{background}{a character specifying the color used for the background (borders).}

\item{cuts}{the number of cuts to make.}

\item{ratio}{the \code{1:1} ratio for each cut.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints a Mondriaan.
}
\examples{
set.seed(6)
paint_mondriaan(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
