% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_planet.R
\name{paint_planet}
\alias{paint_planet}
\title{Paint a Planet on a Canvas}
\usage{
paint_planet <- function(colors, threshold = 3, iterations = 10, starprob = 0.00001,
                     radius = NULL, center.x = NULL, center.y = NULL, 
                     seed = 1, width = 1500, height = 1500)
}
\arguments{
\item{colors}{a character specifying the colors used for the planets}

\item{threshold}{a character specifying the threshold for a color take.}

\item{iterations}{the number of iterations of the planets}

\item{starprob}{the probability of drawing a star in outer space.}

\item{radius}{a numeric (vector) specifying the radius of the planet(s).}

\item{center.x}{the x-axis coordinate(s) for the center(s) of the planet(s).}

\item{center.y}{the y-axis coordinate(s) for the center(s) of the planet(s).}

\item{seed}{the seed for the painting.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints one or multiple planets.
}
\examples{
paint_planet(colors = c("dodgerblue", "forestgreen"))

}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_function}} \code{\link{paint_ant}} \code{\link{paint_mondriaan}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
