% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_polylines.R
\name{paint_polylines}
\alias{paint_polylines}
\title{Paint Polygons and Lines on Canvas}
\usage{
paint_polylines(colors, background = '#fafafa', ratio = 0.5, iterations = 1000, 
                alpha = NULL, size = 0.25, width = 500, height = 500)
}
\arguments{
\item{colors}{a character (vector) specifying the colors used for the strokes.}

\item{background}{a character specifying the color used for the background (borders).}

\item{ratio}{width of the polygons. Larger ratios cause more overlap.}

\item{iterations}{the number of iterations on the painting.}

\item{alpha}{transparency of the polygons.}

\item{size}{size of the borders.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function creates polygons with lines.
}
\examples{
set.seed(1)
paint_polylines(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
