% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_shape.R
\name{paint_shape}
\alias{paint_shape}
\title{Paint shapes}
\usage{
paint_shape(color = '#000000', background = '#fafafa', seed = 1)
}
\arguments{
\item{color}{the color of the shape.}

\item{background}{the color of the background.}

\item{seed}{the seed for the painting.}
}
\value{
A \code{ggplot} object with the painting.
}
\description{
This function paints shapes.
}
\examples{
bg <- sample(c('#fafafa', '#cc7722', '#a9d2c3', '#fc7c7c'), size = 1)

paint_shape(color = '#000000', background = bg)

}
\references{
https://github.com/cutterkom/generativeart
}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_turmite}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
