% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_strokes.R
\name{paint_strokes}
\alias{paint_strokes}
\title{Paint Strokes on a Canvas}
\usage{
paint_strokes(colors = '#000000', neighbors = 1, p = 0.01, seed = 1, 
                     iterations = 1, width = 500, height = 500, side = FALSE)
}
\arguments{
\item{colors}{a vector of colors for the painting.}

\item{neighbors}{the number of neighbors a block considers when taking over a color.}

\item{p}{the probability of selecting a new color at each block.}

\item{seed}{the seed for the painting.}

\item{iterations}{the number of iterations on the painting.}

\item{width}{the width of the painting.}

\item{height}{the height of the painting.}

\item{side}{whether to turn the painting on its side.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function creates a painting that resembles paints strokes. The algorithm is based on the simple idea that each next point on the grid has a chance to take over the color of an adjacent colored point but also has a change of generating a new color.
}
\examples{
paint_strokes(colors = c('#fafafa', '#000000'), neighbors = 1, p = 0.01,
              seed = 1, side = FALSE, iterations = 1, 
              width = 1500, height = 1500)

}
\seealso{
\code{\link{paint_turmite}} \code{\link{paint_function}} \code{\link{paint_ant}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
