% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_turmite.R
\name{paint_turmite}
\alias{paint_turmite}
\title{Paint turmites}
\usage{
paint_turmite(color = '#fafafa', background = '#000000', p = 0.5, seed = 1, 
                     iterations = 1e7, width = 1500, height = 1500)
}
\arguments{
\item{color}{the color of the turmite.}

\item{background}{the color of the background.}

\item{p}{the probability of a state switch within the turmite.}

\item{seed}{the seed for the painting.}

\item{iterations}{the number of iterations of the turmite.}

\item{width}{the width of the painting.}

\item{height}{the height of the painting.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints turmites. A turmite is a Turing machine which has an orientation in addition to a current state and a "tape" that consists of a two-dimensional grid of cells. The algorithm is simple: 1) turn on the spot (left, right, up, down) 2) change the color of the square 3) move forward one square.
}
\examples{
paint_turmite(color = "#fafafa", background = "#1E90FF", p = 0.5,
              seed = 1, iterations = 1e7, width = 1500, height = 1500)

}
\references{
\url{https://en.wikipedia.org/wiki/Turmite}
}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_shape}} \code{\link{paint_ant}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
