% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_x.R
\name{paint_x}
\alias{paint_x}
\title{Paint Strokes on a Canvas}
\usage{
paint_x(colors, background = '#fafafa', ratio = 0.1, iterations = 500, 
        alpha = 0.75, size = 0.5, width = 500, height = 500)
}
\arguments{
\item{colors}{a character (vector) specifying the colors used for the strokes.}

\item{iterations}{the number of iterations on the painting.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function creates a painting that resembles paints strokes. The algorithm is based on the simple idea that each next point on the grid has a chance to take over the color of an adjacent colored point but also has a change of generating a new color.
}
\examples{
set.seed(1)
paint_x(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
