% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCanvas.R
\name{saveCanvas}
\alias{saveCanvas}
\title{Save a canvas to an external device.}
\usage{
saveCanvas(plot, filename, resolution)
}
\arguments{
\item{plot}{a ggplot2 object to be saved.}

\item{filename}{the filename of the export.}

\item{resolution}{the \code{dpi} of the export.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function is a wrapper around \code{ggplot2::ggsave}. It provides a suggested export with square dimensions for a canvas created using the \code{aRtsy} package.
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{canvas}
\keyword{save}
