% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_canvas.R
\name{theme_canvas}
\alias{theme_canvas}
\title{Canvas Theme for ggplot2 Objects}
\usage{
theme_canvas(x, background = '#fafafa', margin = -1.25)
}
\arguments{
\item{x}{a ggplot2 object.}

\item{background}{a character specifying the color used for the empty canvas.}

\item{margin}{margins of the plot.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
Add a canvas theme to the plot. The canvas theme by default has no margins and fills any empty canvas with a background color.
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{canvas}
\keyword{theme}
