% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeCanvas.R
\name{themeCanvas}
\alias{themeCanvas}
\title{Canvas theme for ggplot2 objects}
\usage{
themeCanvas(x)
}
\arguments{
\item{x}{a ggplot2 object.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
Add a canvas theme to the plot. The canvas theme by default has no margins and fills any empty canvas with a background color.
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{canvas}
\keyword{theme}
