% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_ant.R
\name{canvas_ant}
\alias{canvas_ant}
\title{Paint Langton's Ant on a Canvas}
\usage{
canvas_ant(colors, background = '#fafafa', iterations = 1e7,
           width = 200, height = 200)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the ant.}

\item{background}{a character specifying the color of the background.}

\item{iterations}{the number of iterations of the ant.}

\item{width}{the width of the artwork in pixels.}

\item{height}{the height of the artwork in pixels.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function paints Langton's Ant. Langton's ant is a two-dimensional universal Turing machine with a very simple set of rules but complex emergent behavior.
}
\examples{
canvas_ant(colors = '#000000', background = '#fafafa')

}
\references{
\url{https://en.wikipedia.org/wiki/Langton\%27s_ant}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
