% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_blacklight.R
\name{canvas_blacklight}
\alias{canvas_blacklight}
\title{Draw Blacklights}
\usage{
canvas_blacklight(colors, n = 1000, width = 500, height = 500)
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{n}{a positive integer specifying the number of random data points to generate.}

\item{width}{a positive integer specifying the width of the artwork in pixels.}

\item{height}{a positive integer specifying the height of the artwork in pixels.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function draws the predictions from a support vector machines algorithm on randomly generated continuous data.
}
\examples{
\donttest{
set.seed(1)

# Simple example
canvas_blacklight(colors = colorPalette("tuscany2"))
}

}
\references{
\url{https://en.wikipedia.org/wiki/Support-vector_machine}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
