% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_flame.R
\name{canvas_flame}
\alias{canvas_flame}
\title{Draw a Fractal Flame}
\usage{
canvas_flame(colors, background = "#000000", iterations = 500000,
                      resolution = 1000, span = 2)
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{background}{a character specifying the color used for the background.}

\item{iterations}{a positive integer specifying the number of iterations of the algorithm.}

\item{resolution}{resolution of the artwork in pixels per row/column. Increasing the resolution increases the quality of the artwork but also increases the computation time exponentially.}

\item{span}{a value indicating the width.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function draws fractal flames.
}
\examples{
\donttest{
set.seed(2)

# Simple example
canvas_flame(colors = colorPalette("origami"))
}

}
\references{
\url{https://en.wikipedia.org/wiki/Fractal_flame}
}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
