% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_flame.R
\name{canvas_flame}
\alias{canvas_flame}
\title{Draw a Fractal Flame}
\usage{
canvas_flame(colors, background = "#000000",
                      iterations = 1000000, resolution = 1000,
                      variations = NULL, blend = TRUE, 
                      post = FALSE, final = FALSE, extra = FALSE,
                      verbose = FALSE)
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{background}{a character specifying the color used for the background.}

\item{iterations}{a positive integer specifying the number of iterations of the algorithm.}

\item{resolution}{resolution of the artwork in pixels per row/column. Increasing the resolution increases the quality of the artwork but also increases the computation time exponentially.}

\item{variations}{an integer (vector) specifying the variations to be included in the flame. The default \code{NULL} includes all variations. See the details section for more information about possible variations.}

\item{blend}{logical. Whether to blend the variations (\code{TRUE}) or pick a unique variation in each iteration (\code{FALSE}).}

\item{post}{logical. Whether to apply a post transformation in each iteration.}

\item{final}{logical. Whether to apply a final transformation in each iteration.}

\item{extra}{logical. Whether to apply an additional post transformation after the final transformation. Only has an effect when \code{final = TRUE}.}

\item{verbose}{logical. Whether to print information.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function implements the fractal flame algorithm.
}
\details{
The \code{variation} argument can be used to include specific variations into the flame. See the appendix in the references for examples of all variations. Possible variations are:

\itemize{
 \item{\code{0}: Linear}
 \item{\code{1}: Sine}
 \item{\code{2}: Spherical}
 \item{\code{3}: Swirl}
 \item{\code{4}: Horsehoe}
 \item{\code{5}: Polar}
 \item{\code{6}: Handkerchief}
 \item{\code{7}: Heart}
 \item{\code{8}: Disc}
 \item{\code{9}: Spiral}
 \item{\code{10}: Hyperbolic}
 \item{\code{11}: Diamond}
 \item{\code{12}: Ex}
 \item{\code{13}: Julia}
 \item{\code{14}: Bent}
 \item{\code{15}: Waves}
 \item{\code{16}: Fisheye}
 \item{\code{17}: Popcorn}
}
}
\examples{
\donttest{
set.seed(2)

# Simple example
canvas_flame(colors = colorPalette("origami"))
}

}
\references{
\url{https://flam3.com/flame_draves.pdf}
}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
