% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_flows.R
\name{canvas_flows}
\alias{canvas_flows}
\title{Paint Flows}
\usage{
canvas_flows(colors, background = "#fafafa", lines = 100, iterations = 100,
             angles = c("svm", "knn", "rf"), width = 100, height = 100)
}
\arguments{
\item{colors}{a string specifying the color used for the artwork.}

\item{background}{a character specifying the color used for the background.}

\item{lines}{the number of lines to draw.}

\item{iterations}{the maximum number of iterations for each line.}

\item{angles}{method of setting the angles of the flow field. Possible options are \code{knn} and \code{svm}.}

\item{width}{a positive integer specifying the width of the artwork in pixels.}

\item{height}{a positive integer specifying the height of the artwork in pixels.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function draws flow fields on a canvas.
}
\examples{
\donttest{
set.seed(1)

# Simple example
canvas_flows(colors = colorPalette("dark2"))
}

}
\references{
\url{https://tylerxhobbs.com/essays/2020/flow-fields}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
