% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_forest.R
\name{canvas_forest}
\alias{canvas_forest}
\title{Paint a forest on a canvas}
\usage{
canvas_forest(colors, n = 1000, resolution = 500)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the artwork.}

\item{n}{number of data points to generate.}

\item{resolution}{the number of pixels (width and height) of the artwork.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function creates an artwork from randomly generated data by running a random forest classification algorithm to predict the color of each pixel on the canvas.
}
\examples{
\donttest{
set.seed(1)
canvas_forest(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))
}

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
