% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_gemstone.R
\name{canvas_gemstone}
\alias{canvas_gemstone}
\title{Draw Gemstones}
\usage{
canvas_gemstone(colors, n = 1000, width = 500, height = 500)
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{n}{a positive integer specifying the number of random data points to generate.}

\item{width}{a positive integer specifying the width of the artwork in pixels.}

\item{height}{a positive integer specifying the height of the artwork in pixels.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function creates an artwork from randomly generated data by running a k-nearest neighbors regression algorithm to predict the color of each pixel on the canvas.
}
\examples{
\donttest{
set.seed(1)

# Simple example
canvas_gemstone(colors = colorPalette("dark3"))
}

}
\references{
\url{https://en.wikipedia.org/wiki/K-nearest_neighbors_algorithm}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
