% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_ribbons.R
\name{canvas_ribbons}
\alias{canvas_ribbons}
\title{Paint Ribbons on a Canvas}
\usage{
canvas_ribbons(colors, background = '#fdf5e6', triangle = TRUE)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the ribbons. Colors determine the number of ribbons.}

\item{background}{a character specifying the color of the background.}

\item{triangle}{logical. Whether to draw the triangle that breaks the ribbon polygons.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function paints ribbons and (optionally) a triangle in the middle.
}
\examples{
\donttest{
set.seed(1)
canvas_ribbons(colors = colorPalette('tuscany1'))
}

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
