% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas_smoke.R
\name{canvas_smoke}
\alias{canvas_smoke}
\title{Draw Rainbow Smoke}
\usage{
canvas_smoke(colors, resolution = 150, compute = c("minimum", "average"))
}
\arguments{
\item{colors}{a string or character vector specifying the color(s) used for the artwork.}

\item{resolution}{resolution of the artwork in pixels per row/column. Increasing the resolution increases the quality of the artwork but also increases the computation time exponentially.}

\item{compute}{a character specifying how to select a new pixel. The default option \code{minimum} selects the pixel with the smallest color difference in a single neighbor and is relatively fast. The option \code{average} selects the pixel with the smallest average color difference in all the neighbors and is relatively slow.}
}
\value{
A \code{ggplot} object containing the artwork.
}
\description{
This function implements the rainbow smoke algorithm.
}
\examples{
\donttest{
set.seed(1)

# Simple example
canvas_smoke(colors = colorPalette("random", 1024), compute = "average")

# Advanced example
canvas_smoke(colors = "all", resolution = 500)
}

}
\references{
\url{http://rainbowsmoke.hu}
}
\seealso{
\code{colorPalette}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{artwork}
\keyword{canvas}
