% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_ant.R
\name{paint_ant}
\alias{paint_ant}
\title{Paint Langton's Ant on a Canvas}
\usage{
paint_ant(colors, background = '#fafafa', iterations = 1e7,
          seed = 1, width = 200, height = 200)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the ant.}

\item{background}{a character specifying the color of the background.}

\item{iterations}{the number of iterations of the ant.}

\item{seed}{the seed for the painting.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints Langton's Ant. Langton's ant is a two-dimensional universal Turing machine with a very simple set of rules but complex emergent behavior.
}
\examples{
paint_ant(colors = '#000000', background = '#fafafa')

}
\references{
\url{https://en.wikipedia.org/wiki/Langton\%27s_ant}
}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_function}} \code{\link{paint_turmite}} \code{\link{paint_mondriaan}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
