% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_arcs.R
\name{paint_arcs}
\alias{paint_arcs}
\title{Paint Arcs on a Canvas}
\usage{
paint_arcs(colors, background = '#fafafa', n = 9, nrow = NULL, ncol = NULL, seed = 1)
}
\arguments{
\item{background}{a character specifying the color used for the background.}

\item{seed}{the seed for the painting.}

\item{color}{a character specifying the three colors used for the painting.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints arcs.
}
\examples{
paint_arcs(colors = c('black', 'red', 'yellow'))

}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_turmite}} \code{\link{paint_ant}} \code{\link{paint_mondriaan}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
