% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_function.R
\name{paint_function}
\alias{paint_function}
\title{Paint Functions on a Canvas}
\usage{
paint_function(color, background = '#fafafa')
}
\arguments{
\item{color}{a character specifying the color used for the function shape.}

\item{background}{a character specifying the color used for the background.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints functions with random parameters and mimics the functionality of the \code{generativeart} package.
}
\examples{
set.seed(1)
paint_function(color = '#000000', background = '#fafafa')

}
\references{
\url{https://github.com/cutterkom/generativeart}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
