% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_mondriaan.R
\name{paint_mondriaan}
\alias{paint_mondriaan}
\title{Paint a Mondriaan on a Canvas}
\usage{
paint_mondriaan(colors = '#fafafa', background = '#000000', cuts = 50,
                ratio = 1.618, seed = 1, width = 200, height = 200)
}
\arguments{
\item{colors}{the colors of the squares.}

\item{background}{the color of the background (borders).}

\item{cuts}{the number of cuts to make.}

\item{ratio}{the \code{1:1} ratio for each cut.}

\item{seed}{the seed for the painting.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints a Mondriaan.
}
\examples{
paint_mondriaan(colors = c('white', 'yellow', 'red'), 
                background = '#000000', cuts = 20, ratio = 1.618, 
                seed = 12, width = 200, height = 200)

}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_turmite}} \code{\link{paint_ant}} \code{\link{paint_function}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
