% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_ribbons.R
\name{paint_ribbons}
\alias{paint_ribbons}
\title{Paint Ribbons on a Canvas}
\usage{
paint_ribbons(colors, background = '#fdf5e6', triangle = TRUE)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the ribbons.}

\item{background}{a character specifying the color of the background.}

\item{triangle}{logical. Whether to draw the triangle that breaks the ribbon polygons.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints ribbons and (optionally) a triangle in the middle.
}
\examples{
paint_ribbons(colors = c("forestgreen", "firebrick", "dodgerblue", "goldenrod"))

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
