% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_triangle.R
\name{paint_ribbons}
\alias{paint_ribbons}
\title{Paint A Triangle with Lines on a Canvas}
\usage{
paint_ribbons(colors, background = '#fdf5e6', seed = 1)
}
\arguments{
\item{colors}{a character (vector) specifying the colors for the lines.}

\item{background}{a character specifying the color of the background.}

\item{triangle}{logical. Whether to draw the triangle itself.}

\item{seed}{the seed for the painting.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function paints triangles and lines.
}
\examples{
paint_ribbons(colors = c("forestgreen", "firebrick", "dodgerblue", "goldenrod"))

}
\seealso{
\code{\link{paint_strokes}} \code{\link{paint_turmite}} \code{\link{paint_ant}} \code{\link{paint_mondriaan}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
