% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_strokes.R
\name{paint_strokes}
\alias{paint_strokes}
\title{Paint strokes}
\usage{
paint_strokes(palette = '#000000', neighbors = 1, p = 0.01, seed = 1, 
                     iter = 1, width = 500, height = 500)
}
\arguments{
\item{palette}{a vector of colors for the painting.}

\item{neighbors}{the number of neighbors a block considers when taking over a color.}

\item{p}{the probability of .}

\item{seed}{the seed for the painting.}

\item{iter}{the number of iterations.}

\item{width}{the width of the painting.}

\item{height}{the height of the painting.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function creates a painting that resembles paints strokes. The algorithm is based on the simple idea that each next point on the grid has a chance to take over the color of an adjacent colored point but also has a change of generating a new color.
}
\examples{
paint_strokes(palette = c('#fafafa', '#000000'), neighbors = 1, p = 0.01,
              seed = 1, iter = 1, width = 1500, height = 1500)

}
\seealso{
\code{\link{paint_turmite}} \code{\link{paint_shape}}
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
