% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paint_strokes.R
\name{paint_strokes}
\alias{paint_strokes}
\title{Paint Strokes on a Canvas}
\usage{
paint_strokes(colors, neighbors = 1, p = 0.01, iterations = 1, 
                     seed = 1, width = 500, height = 500, side = FALSE)
}
\arguments{
\item{colors}{a character (vector) specifying the colors used for the strokes.}

\item{neighbors}{the number of neighbors a block considers when taking over a color. More neighbors fades the painting.}

\item{p}{the probability of selecting a new color at each block. A higher probability adds more noise to the painting.}

\item{iterations}{the number of iterations on the painting. More iterations fade the painting.}

\item{seed}{the seed for the painting.}

\item{width}{the width of the painting in pixels.}

\item{height}{the height of the painting in pixels.}

\item{side}{whether to turn the painting on its side.}
}
\value{
A \code{ggplot} object containing the painting.
}
\description{
This function creates a painting that resembles paints strokes. The algorithm is based on the simple idea that each next point on the grid has a chance to take over the color of an adjacent colored point but also has a change of generating a new color.
}
\examples{
paint_strokes(colors = c('forestgreen', 'goldenrod', 'firebrick', 'navyblue'))

}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{paint}
