% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCanvas.R
\name{saveCanvas}
\alias{saveCanvas}
\title{Save a Canvas to an External Device}
\usage{
saveCanvas(plot, filename, width = 7, height = 7, dpi = 300)
}
\arguments{
\item{plot}{a ggplot2 object to be saved.}

\item{filename}{the filename of the export.}

\item{width}{the width of the artwork in cm.}

\item{height}{the height of the artwork in cm.}

\item{dpi}{the \code{dpi} (dots per inch) of the file.}
}
\value{
No return value, called for saving plots.
}
\description{
This function is a wrapper around \code{ggplot2::ggsave}. It provides a suggested export with square dimensions for a canvas created using the \code{aRtsy} package.
}
\author{
Koen Derks, \email{koen-derks@hotmail.com}
}
\keyword{canvas}
\keyword{save}
